Extract window api impls
